requirements.txt
absl-py==0.8.1
appnope==0.1.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
fastcache==1.1.0
graphviz==0.13.2
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jax==0.1.49
jaxlib==0.1.32
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.17.3
numpyro==0.2.1
opt-einsum==3.1.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.4
pyreadr==0.2.1
pyro-api==0.1.1
pyro-ppl==1.0.0
pyrsistent==0.15.5
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.0
qtconsole==4.5.5
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.13.0
sklearn==0.0
statsmodels==0.10.1
terminado==0.8.2
testpath==0.4.4
torch==1.3.1
tornado==6.0.3
tqdm==4.37.0
traitlets==4.3.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==0.6.0
